pyproject.toml
[project]
name = "crypto-lakehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.12.2",
    "delta-spark==3.3.0",
    "deltalake>=1.3.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyspark==3.5.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
